requirements.txt
beautifulsoup4==4.8.1
bs4==0.0.1
cycler==0.10.0
dateparser==0.7.2
dill==0.3.1.1
distro==1.5.0
edlib==1.3.8.post1
joblib==0.14.0
kiwisolver==1.1.0
lxml==4.4.2
matplotlib==3.1.2
natsort==6.2.0
nltk==3.4.5
numpy==1.17.4
pandas==1.0.4
pdf2image==1.14.0
Pillow==6.2.1
pycairo==1.19.0
PyGObject==3.34.0
PyMySQL==0.9.3
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
regex==2019.11.1
scikit-learn==0.22
scipy==1.3.3
seaborn==0.10.1
six==1.13.0
sklearn==0.0
soupsieve==1.9.5
tabula-py==2.1.1
tabulate==0.8.7
torch
torchvision
tzlocal==2.0.0
Wand==0.5.7
